% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_weights.R
\name{add_weights_list}
\alias{add_weights_list}
\title{Compare various methods to a simulated data set}
\usage{
add_weights_list(weights_list, risk_table)
}
\description{
\code{compare_weights} will simulate a data set (according to \code{design})
and will calculate standardized z-statistics for a variety of methods
\param{weights_list_list} A list, where each element is a list describing a method. For example,
\code{list(method="fixed_c",delay=6)} or \code{list(method="landmark", time = 20)}.
\param{design} A list of design information, containing:
\code{med_c}, \code{rate_e_1}, \code{rate_e_2}, \code{rec_period}, \code{rec_power}, \code{delay}, \code{max_cal_t}
\return A vector of standardized test statistics, one for each method in \code{weights_list_list}.
}
