% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_weights.R
\name{compare_weights}
\alias{compare_weights}
\title{Compare various methods to a simulated data set}
\usage{
compare_weights(weights_list_list, design)
}
\arguments{
\item{{weights_list_list}}{A list, where each element is a list describing a method. For example,
\code{list(method="fixed_c",delay=6)} or \code{list(method="landmark", time = 20)}.}

\item{{design}}{A list of design information, containing:
\code{med_c}, \code{rate_e_1}, \code{rate_e_2}, \code{rec_period}, \code{rec_power}, \code{delay}, \code{max_cal_t}}
}
\value{
A vector of standardized test statistics, one for each method in \code{weights_list_list}.
}
\description{
\code{compare_weights} will simulate a data set (according to \code{design})
and will calculate standardized z-statistics for a variety of methods
}
