% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delayed_effect_sim.R
\name{delayed_effect_sim}
\alias{delayed_effect_sim}
\title{Simulate survival data from a two-arm trial}
\usage{
delayed_effect_sim(n_c = 100, n_e = 100, rec_period = 12,
  rec_power = 1, med_c = 15, rate_e_1 = log(2)/9, rate_e_2 = 0.04,
  delay = 6, max_cal_t = 36)
}
\arguments{
\item{{n_c}}{Number of patients on control treatment.}

\item{{n_e}}{Number of patients on experimental treatment.}

\item{{rec_period}}{Recruitment period.}

\item{{rec_power}}{Recruitment follows a power model. Pr(recruited before T) = (T / rec_period) ^ rec_power.}

\item{{med_c}}{Median survival time on control.}

\item{{rate_e_1}}{Event rate during first period on experimental arm.}

\item{{rate_e_2}}{Event rate during second period on experimental arm.}

\item{{delay}}{Length of first period.}

\item{{max_cal_t}}{Maximum calendar time, i.e., time from start of the trial to data cut-off.}
}
\value{
A data frame containing survival time, whether patient has event (1 = yes, 0 = censored), and treatment arm.
}
\description{
\code{delayed_effect_sim} Simulate survival data: exponential vs. two-piece exponential.
}
