% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_df.R
\name{get_df}
\alias{get_df}
\title{Simulate survival data from a two-arm trial}
\usage{
get_df(n_c = 100, n_e = 100, rec_period = 12, rec_power = 1,
  med_c = 15, rate_e_1 = log(2)/9, rate_e_2 = 0.04, delay = 6,
  max_cal_t = 36)
}
\description{
\code{get_df_strong} Simulate survival data: exponential vs. two-piece exponential.
\param{n_c} Number of patients on control treatment.
\param{n_e} Number of patients on experimental treatment.
\param{rec_period} Recruitment period.
\param{rec_power} Recruitment follows a power model. Pr(recruited before T) = (T / rec_period) ^ rec_power.
\param{med_c} Median survival time on control.
\param{rate_e_1} Event rate during first period on experimental arm.
\param{rate_e_2} Event rate during second period on experimental arm.
\param{delay} Length of first period.
\param{max_cal_t} Maximum calendar time, i.e., time from start of the trial to data cut-off.
\return A data frame containing survival time, whether patient has event (1 = yes, 0 = censored), and treatment arm.
}
