% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncp_power.R
\name{ncp_power}
\alias{ncp_power}
\title{Find the approximate non-centrality parameter and power of a modestWLRT.}
\usage{
ncp_power(t_star, model, recruitment, dco, length_t,
  alpha_one_sided = 0.025)
}
\arguments{
\item{{t_star}}{A vector. A range of possible values for t*.}

\item{{model}}{A piecewise constant hazard model.
A list containing the \code{change_points}; the rates \code{lambdas_0} on the control arm; 
and the rates \code{lambdas_1} on the treatment arm.}

\item{{recruitment}}{List of recruitment information. 
Containing \enumerate{
              \item Sample size on control, \code{n_0} 
              \item Sample size on treatment, \code{n_1} 
              \item Recruitment period, \code{r_period}
              \item Recruitment parameter for power model, \code{k} 
            }}

\item{{dco}}{Time of data cut-off.}

\item{{length_t}}{}

\item{{alpha_one_sided}}{}
}
\value{
A list containing
\enumerate{
                \item non-centrality parameter corresponding to each t* \code{ncp} 
                \item power corresponding to each t* \code{power}
          }
}
\description{
\code{ncp_power} returns the approximate non-centrality parameter and power of a modestWLRT for a range
of possible values of t*.
}
