% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmst.R
\name{rmst}
\alias{rmst}
\title{Perform a restricted-mean survival time.}
\usage{
rmst(df, time)
}
\arguments{
\item{{time}}{The (patient) time to perform the RMST analysis at. Must be length 1.}
}
\value{
\code{z} the standardized test statistic. Large values indicate better survival on the experimental arm.
}
\description{
\code{rmst} performs a restricted mean survival time analysis at a specified (patient) time.
}
