% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three_piece_sim.R
\name{three_piece_sim}
\alias{three_piece_sim}
\title{Simulate survival data from a two-arm trial}
\usage{
three_piece_sim(n_c = 100, n_e = 100, rec_period = 12,
  rec_power = 1, rate_c_1 = log(2)/9, rate_c_2 = 0.04,
  rate_c_3 = 0.04, rate_e_1 = log(2)/9, rate_e_2 = 0.04,
  rate_e_3 = 0.04, delay_1 = 6, delay_2 = 12, max_cal_t = 36,
  n_events = NULL)
}
\arguments{
\item{{n_c}}{Number of patients on control treatment.}

\item{{n_e}}{Number of patients on experimental treatment.}

\item{{rec_period}}{Recruitment period.}

\item{{rec_power}}{Recruitment follows a power model. Pr(recruited before T) = (T / rec_period) ^ rec_power.}

\item{{rate_c_1}}{Event rate during first period on control arm.}

\item{{rate_c_2}}{Event rate during second period on control arm.}

\item{{rate_c_3}}{Event rate during third period on control arm.}

\item{{rate_e_1}}{Event rate during first period on experimental arm.}

\item{{rate_e_2}}{Event rate during second period on experimental arm.}

\item{{rate_e_3}}{Event rate during third period on experimental arm.}

\item{{delay_1}}{Time of first change point.}

\item{{delay_2}}{Time of second change point.}

\item{{max_cal_t}}{Maximum calendar time, i.e., time from start of the trial to data cut-off.}
}
\value{
A data frame containing survival time, whether patient has event (1 = yes, 0 = censored), and treatment arm.
}
\description{
\code{three_piece_sim} Simulate survival data: three-piece exponential vs. three-piece exponential.
}
